Improved updateLesson and reqlession see description.
When clicking on a lesson to update, you now don't have to remember
which data of a lesson you want to update. The program shows now the
current data of a lesson which you want to update.

In reglesson.php the program checks wether the student is registering
himself/herself on time. If that's the case, then the student is
present. If that's not the case, then the student will not be registered
as present.